+++ config/src/main/java/org/hyperledger/besu/config/JsonGenesisConfigOptions.java
@@ -290,6 +290,11 @@ public class JsonGenesisConfigOptions implements GenesisConfigOptions {
   }
 
   @Override
+  public OptionalLong getEcip1017EraRounds() {
+    return getOptionalLong("ecip1017erarounds");
+  }
+
+  @Override
   public Map<String, Object> asMap() {
     final ImmutableMap.Builder<String, Object> builder = ImmutableMap.builder();
     getChainId().ifPresent(chainId -> builder.put("chainId", chainId));
@@ -322,6 +327,8 @@ public class JsonGenesisConfigOptions implements GenesisConfigOptions {
     getEIP1559BlockNumber().ifPresent(l -> builder.put("eip1559Block", l));
     getContractSizeLimit().ifPresent(l -> builder.put("contractSizeLimit", l));
     getEvmStackSize().ifPresent(l -> builder.put("evmstacksize", l));
+    getEcip1017EraRounds().ifPresent(l -> builder.put("ecip1017EraRounds", l));
+
     if (isClique()) {
       builder.put("clique", getCliqueConfigOptions().asMap());
     }