+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/privacy/PrivateStateRehydration.java
@@ -88,16 +88,26 @@ public class PrivateStateRehydration {
           .updater()
           .putPrivacyGroupHeadBlockMap(
               getBlockHashForIndex(0, privateTransactionWithMetadataList),
-              PrivacyGroupHeadBlockMap.EMPTY)
+              PrivacyGroupHeadBlockMap.empty())
           .commit();
+    }
+
+    final LinkedHashMap<Hash, PrivateTransaction> pmtHashToPrivateTransactionMap =
+        new LinkedHashMap<>();
+    for (int j = 0; j < privateTransactionWithMetadataList.size(); j++) {
+      final PrivateTransactionWithMetadata transactionWithMetadata =
+          privateTransactionWithMetadataList.get(j);
+      pmtHashToPrivateTransactionMap.put(
+          transactionWithMetadata.getPrivateTransactionMetadata().getPrivacyMarkerTransactionHash(),
+          transactionWithMetadata.getPrivateTransaction());
     }
 
     for (int i = 0; i < privateTransactionWithMetadataList.size(); i++) {
       // find out which block this transaction is in
       final Hash blockHash = getBlockHashForIndex(i, privateTransactionWithMetadataList);
 
-      // if there are multiple pmts in the list we can increment our index i. At the end of the
-      // while loop i will be the index of the last PMT (for this group) that is in this block.
+      // At the end of the while loop i will be the index of the last PMT (for this group) that is
+      // in this block.
       while (i + 1 < privateTransactionWithMetadataList.size()
           && blockHash.equals(getBlockHashForIndex(i + 1, privateTransactionWithMetadataList))) {
         i++;
@@ -127,8 +137,7 @@ public class PrivateStateRehydration {
               .map(Transaction::getHash)
               .collect(Collectors.toList()));
 
-      final ProtocolSpec protocolSpec =
-          protocolSchedule.getByBlockNumber(blockchain.getBlockHeader(blockHash).get().getNumber());
+      final ProtocolSpec protocolSpec = protocolSchedule.getByBlockNumber(blockHeader.getNumber());
       final PrivateGroupRehydrationBlockProcessor privateGroupRehydrationBlockProcessor =
           new PrivateGroupRehydrationBlockProcessor(
               protocolSpec.getTransactionProcessor(),
@@ -145,18 +154,6 @@ public class PrivateStateRehydration {
               .flatMap(publicWorldStateArchive::getMutable)
               .orElseThrow(RuntimeException::new);
 
-      // enclave cache for private block rehydration
-      final LinkedHashMap<Hash, PrivateTransaction> enclaveMap = new LinkedHashMap<>();
-      for (int j = 0; j < privateTransactionWithMetadataList.size(); j++) {
-        final PrivateTransactionWithMetadata transactionWithMetadata =
-            privateTransactionWithMetadataList.get(j);
-        enclaveMap.put(
-            transactionWithMetadata
-                .getPrivateTransactionMetadata()
-                .getPrivacyMarkerTransactionHash(),
-            transactionWithMetadata.getPrivateTransaction());
-      }
-
       privateGroupRehydrationBlockProcessor.processBlock(
           blockchain,
           publicWorldState,
@@ -164,7 +161,7 @@ public class PrivateStateRehydration {
           privateStateStorage,
           privateStateRootResolver,
           block,
-          enclaveMap,
+          pmtHashToPrivateTransactionMap,
           block.getBody().getOmmers());
 
       // check the resulting private state against the state in the meta data
@@ -217,7 +214,7 @@ public class PrivateStateRehydration {
       final PrivacyGroupHeadBlockMap thePrivacyGroupHeadBlockMap =
           privateStateStorage
               .getPrivacyGroupHeadBlockMap(theBlockHeader.getHash())
-              .orElse(PrivacyGroupHeadBlockMap.EMPTY);
+              .orElse(PrivacyGroupHeadBlockMap.empty());
       final PrivateStateStorage.Updater privateStateUpdater = privateStateStorage.updater();
       thePrivacyGroupHeadBlockMap.put(privacyGroupId, hashOfLastBlockWithPmt);
       privateStateUpdater.putPrivacyGroupHeadBlockMap(