+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/vm/operations/CodeSizeOperation.java
@@ -28,11 +28,7 @@ public class CodeSizeOperation extends AbstractFixedCostOperation {
   }
 
   @Override
-  public OperationResult execute(final MessageFrame frame, final EVM evm) {
-    if (frame.getRemainingGas().compareTo(gasCost) < 0) {
-      return outOfGasResponse;
-    }
-
+  public OperationResult executeFixedCostOperation(final MessageFrame frame, final EVM evm) {
     final Code code = frame.getCode();
     frame.pushStackItem(UInt256.valueOf(code.getSize()).toBytes());