+++ privacy-contracts/src/main/solidity/OnChainPrivacyGroupManagementInterface.sol
@@ -1,12 +1,25 @@
-pragma solidity ^0.5.9;
-
+/*
+ * Copyright ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+pragma solidity ^0.6.0;
 interface OnChainPrivacyGroupManagementInterface {
 
-    function addParticipants(bytes32 enclaveKey, bytes32[] calldata participants) external returns (bool);
+    function addParticipants(bytes32[] calldata publicEnclaveKeys) external returns (bool);
 
-    function removeParticipant(bytes32 enclaveKey, bytes32 account) external returns (bool);
+    function removeParticipant(bytes32 participant) external returns (bool);
 
-    function getParticipants(bytes32 enclaveKey) external view returns (bytes32[] memory);
+    function getParticipants() external view returns (bytes32[] memory);
 
     function lock() external;
 
@@ -15,4 +28,6 @@ interface OnChainPrivacyGroupManagementInterface {
     function canExecute() external view returns (bool);
 
     function getVersion() external view returns (bytes32);
+
+    function canUpgrade() external returns (bool);
 }