+++ privacy-contracts/src/main/solidity/OnChainPrivacyGroupManagementProxy.sol
@@ -1,5 +1,18 @@
-pragma solidity ^0.5.12;
-
+/*
+ * Copyright ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+pragma solidity ^0.6.0;
 import "./OnChainPrivacyGroupManagementInterface.sol";
 
 contract OnChainPrivacyGroupManagementProxy is OnChainPrivacyGroupManagementInterface {
@@ -10,47 +23,67 @@ contract OnChainPrivacyGroupManagementProxy is OnChainPrivacyGroupManagementInte
         implementation = _implementation;
     }
 
-    function upgradeTo(address _newImplementation) external {
-        require(implementation != _newImplementation);
-        _setImplementation(_newImplementation);
-    }
-
     function _setImplementation(address _newImp) internal {
         implementation = _newImp;
     }
 
-    function addParticipants(bytes32 enclaveKey, bytes32[] memory participants) public returns (bool) {
+    function addParticipants(bytes32[] memory _publicEnclaveKeys) public override returns (bool) {
         OnChainPrivacyGroupManagementInterface privacyInterface = OnChainPrivacyGroupManagementInterface(implementation);
-        return privacyInterface.addParticipants(enclaveKey, participants);
+        return privacyInterface.addParticipants(_publicEnclaveKeys);
     }
 
-    function getParticipants(bytes32 enclaveKey) view public returns (bytes32[] memory) {
+    function getParticipants() view public override returns (bytes32[] memory) {
         OnChainPrivacyGroupManagementInterface privacyInterface = OnChainPrivacyGroupManagementInterface(implementation);
-        return privacyInterface.getParticipants(enclaveKey);
+        return privacyInterface.getParticipants();
     }
 
-    function removeParticipant(bytes32 enclaveKey, bytes32 account) public returns (bool) {
+    function removeParticipant(bytes32 _participant) public override returns (bool) {
         OnChainPrivacyGroupManagementInterface privacyInterface = OnChainPrivacyGroupManagementInterface(implementation);
-        return privacyInterface.removeParticipant(enclaveKey, account);
+        bool result = privacyInterface.removeParticipant(_participant);
+        if (result) {
+            emit ParticipantRemoved(_participant);
+        }
+    return result;
     }
 
-    function lock() public {
+    function lock() public override {
         OnChainPrivacyGroupManagementInterface privacyInterface = OnChainPrivacyGroupManagementInterface(implementation);
         return privacyInterface.lock();
     }
 
-    function unlock() public {
+    function unlock() public override {
         OnChainPrivacyGroupManagementInterface privacyInterface = OnChainPrivacyGroupManagementInterface(implementation);
         return privacyInterface.unlock();
     }
 
-    function canExecute() public view returns (bool) {
+    function canExecute() public view override returns (bool) {
         OnChainPrivacyGroupManagementInterface privacyInterface = OnChainPrivacyGroupManagementInterface(implementation);
         return privacyInterface.canExecute();
     }
 
-    function getVersion() public view returns (bytes32) {
+    function getVersion() public view override returns (bytes32) {
         OnChainPrivacyGroupManagementInterface privacyInterface = OnChainPrivacyGroupManagementInterface(implementation);
         return privacyInterface.getVersion();
     }
+
+    function canUpgrade() external override returns (bool) {
+        OnChainPrivacyGroupManagementInterface privacyInterface = OnChainPrivacyGroupManagementInterface(implementation);
+        return privacyInterface.canUpgrade();
+    }
+
+    function upgradeTo(address _newImplementation) external {
+        require(this.canExecute(), "The contract is locked.");
+        require(implementation != _newImplementation, "The contract to upgrade to has to be different from the current management contract.");
+        require(this.canUpgrade(), "Not allowed to upgrade the management contract.");
+        bytes32[] memory participants = this.getParticipants();
+        _setImplementation(_newImplementation);
+        OnChainPrivacyGroupManagementInterface privacyInterface = OnChainPrivacyGroupManagementInterface(implementation);
+        privacyInterface.addParticipants(participants);
+    }
+
+    event ParticipantRemoved(
+        bytes32 publicEnclaveKey
+    );
+
+
 }